Disable BlockBreak_NoSwing
The "new" minecraft shows the block breaking animation to all players
now. I guess thats enough to distinguish the player that is trying to
break a block from your area. Also the "Direction" and "Visible" check
will force the player to look at the block that he/she/it wants to
break.